requirements.txt
# Database Migration Tool - Requirements
# SQL Database Drivers
pymysql==1.1.0
pyodbc==5.0.1
sqlalchemy==2.0.23

# MongoDB Driver
pymongo==4.6.0

# Configuration and Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Reporting
colorlog==6.8.0

# Data Processing
pandas==2.1.4

